cropping + uploading probably (almost) works

+++ www/js/picHandling.js
@@ -4,43 +4,48 @@ $( document ).ready(function() {
 	$(document).on('change',  function(e) {
 		var target = e.target;
 		if ($(e.target).attr('type') == 'file') {
-			var pic = $(target).parent().prev();
-			var picFile = URL.createObjectURL(e.target.files[0]);
-			pic.attr('src', picFile);
-			console.log("putting new pic in");
-			var img = new Image();
-			img.onload = function() {
-				var data = {};
-				data.origX = 0
-				data.origY = 0;
-				data.origW = this.width;
-				data.origH = this.height;
-				data.src = picFile;
+			var files = e.target.files;
+			console.log("Files is: ", files);
+			if (files.length > 0) {
+				console.log("setting stuff");
+				var picFile = URL.createObjectURL(files[0]);
+				var img = new Image();
+				img.onload = function() {
+					var data = {};
+					data.origX = 0
+					data.origY = 0;
+					data.origW = this.width;
+					data.origH = this.height;
+					data.src = picFile;
 
-				//Save data object in array
-				console.log("right before index");
-				console.log(document.getElementById("testPic").getAttribute("picindex"));
+					//Save data object in array
+					picIndex = parseInt($(e.target).parent().prev().children().attr("picindex"));
+					console.log("just set picIndex to ", picIndex);
+					picData[picIndex] = data;
 
-				picIndex = parseInt(document.getElementById("testPic").getAttribute("picindex"));
-				console.log("JUST SET PICINDEX TO ", picIndex);
-				console.log("Index: ", picIndex);
-				console.log("after index");
-				picData[picIndex] = data;
+				}
+				img.src = picFile;
 
+				if(jcrop_api) {
+					jcrop_api.destroy();
+				}
+				console.log("pic  to set");
+				console.log($(e.target));
+				console.log($(e.target).parent());
+				console.log($(e.target).parent().prev());
+				console.log($(e.target).parent().prev().children());
+
+				$(e.target).parent().prev().children().attr('src', picFile);
+				$("#pic-to-crop").attr('src', picFile);
+				setUpPhotoCrop(name);
+				jcrop_api.setImage(picFile); 
+				$( "#crop-popup" ).popup( "open" );
+			} else {
+				//Clear input
+				$(e.target).val('');
+				//Clear pic
+				$(e.target).parent().prev().children().attr('src', '');
 			}
-			img.src = picFile;
-
-			if(jcrop_api) {
-				jcrop_api.destroy();
-			}
-
-			$("#testPic").attr('src', picFile);
-			$("#pic-to-crop").attr('src', picFile);
-			setUpPhotoCrop(name);
-			jcrop_api.setImage(picFile); 
-			console.log("data set");
-			
-			$( "#test-popup" ).popup( "open" );
 		}
 	});
 
@@ -48,8 +53,7 @@ $( document ).ready(function() {
 });
 
 
-var picData = [null, null, null]; //TODO: Right number 
-var boundx, boundy;
+var picData = [null, null, null, null]; //TODO: Right number 
 var jcrop_api;
 var picIndex;
 
@@ -58,17 +62,15 @@ jQuery(function($) {
 });
 
 
+function clearFile(name) {
+	$("#" + name + 'Button').val('');
+	$("#" + name + 'Pic').attr('src', '');
+	jcrop_api.destroy();
+}
 
-
-
-$( window ).resize(function() {
-	//Anything window-size-dependent changes here!
-	setUpPhotoCrop();
-});
 
 
 function cropBoxSetup(totalWidth) {
-	console.log("cropBoxSetup");
 	jcrop_api = $.Jcrop('#pic-to-crop', {
 		setSelect:   [ 0, 0, 80, 45 ],
 		aspectRatio: 16 / 9,
@@ -89,68 +91,52 @@ function setUpPhotoCrop(name) {
 
 
 function updatePic(c) {
-	console.log("update pic called");
-	let xsize =  $("#test-container").width();
-	let ysize =  $("#test-container").height();
+	let xsize =  250; //medium-pic-container css width
+	let ysize =  141; //medium-pic-container css height
 	var rx = xsize / c.w;
 	var ry = ysize / c.h;
-	console.log("picIndex: ", picIndex);
-	console.log(picData);
 	var data = picData[picIndex]; //Deep or shallow?
-	console.log("DATA: ", data);
+
 	if (data) {
 		data.finalX = c.x;
 		data.finalY = c.y;
 		data.finalW = c.w;
 		data.finalH = c.H;
 
-		console.log("DATA HERE!", data);
-
-		$("#testPic").css({
+		console.log("PIC INDEX IS ", picIndex);
+		$('[picindex=' + picIndex + ']').css({
 			width: Math.round(rx * data.origW) + 'px',
 			height: Math.round(ry * data.origH) + 'px',
 			marginLeft: '-' + Math.round(rx * c.x) + 'px',
 			marginTop: '-' + Math.round(ry * c.y) + 'px'
 		});
-
 	}
-	
-
-
-	
 }
 
 
 
-function savePic(index) {
-	console.log("about to save pic");
+function savePic(index, picName, userId, serial) {
+	console.log("saving pic");
 	var data = picData[index];
-	console.log("Data: ", data);
 	//Save cropped pic in firebases
 	var canvas = document.getElementById("cropCanvas");
-	console.log("Canvas: ", canvas);
 	var context = canvas.getContext('2d');
-
-
-	console.log("Context: ", context);
 	var newPic = new Image();
+
 	newPic.onload = function() {
-		console.log("in onload");
-		//http://stackoverflow.com/questions/37873808/how-can-i-save-canvas-as-image-to-firebase-storage
+		console.log("about to draw image");
 		context.drawImage(newPic, data.origX, data.origY, data.origW, data.origH, data.finalX, data.finalY, data.finalW, data.finalH);
 		console.log("drew image");
-
 		canvas.toBlob(function(blob) {
-			console.log("to blob");
-			// var croppedPic = new Image();
-			// croppedPic.src = blob;
-			firebase.storage().ref().child('testing').put(blob);
-			console.log("firebase!");
+			console.log("got blob");
+			firebase.storage().ref().child("users/" + userId + "/bikes/" + serial + "/" + picName).put(blob);
 		})
-	}//http://www.html5canvastutorials.com/tutorials/html5-canvas-image-crop/
-	console.log("newpic.src");
-	newPic.src = data.src;
-
+	}
+	if (data.src) {
+		console.log("got a src");
+		newPic.src = data.src;
+	}
+	
 }
